Use reflection to simplify the fetcher signature
This patch updates the fetcher library to use reflection in order to
allow the "fetch" functions to return a strongly typed map, rather than
only `map[int64]interface{}`. While the excessive use of reflection is
less than desirable, it greatly simplifies integration with this
library, removing the boilerplate of translating untyped maps to typed
maps.